# Cleanup SDL `output` handling