--- a/include/sdlmain.h
+++ b/include/sdlmain.h
@@ -178,11 +178,9 @@ struct SDL_Block {
 		int pitch = 0;
 		void* framebuf = nullptr;
 		GLuint texture;
-		GLuint displaylist;
 		GLint max_texsize;
 		bool bilinear;
 		bool npot_textures_supported = false;
-		bool use_shader;
 		bool framebuffer_is_srgb_encoded;
 		GLuint program_object;