~~This PR fulfils the prophecy of surface output meeting its demise in cleansing flames of Purgatory. They will not be missed.~~ Extracted into #2780.

As a side bonus, keymapper was ported to SDL renderer and now reuses the window instead of creating a new one (extracted into #2763). Some functionality was moved into separate routines to make initialization and handling of video mode changes less painful to read.

Some (a lot of) testing and feedback is required. Almost all of `[sdl]` configuration block is potentially affected. While testing all possible combinations is an overkill, ensuring that each setting works with both remaining outputs on every platform would be nice: Windows, macOS, X11, Wayland, KMS/DRM. Also please verify whether HiDPI and multi-monitor still work as expected. One interesting case is to check how `fullresolution = original` behaves on a real CRT, if at all.

If you feel I've missed some comments (or punctuation in them), please let me know. Same for moving more stuff around, checks, etc. I'd like to get feedback, else this work could easily drag for six months more without much progress.

Part of work outlined in #2131.