@FeralChild64 2023-05-03
I have found a regression on a dual-monitor setup, using Linux with X.org - after switching to fullscreen, mouse does not work in seamless mode - one has co capture it using middle button. I'll try to look at it, just not today. Extract from my configuration:

```ini
[sdl]
fullscreen          = false
display             = 0
fullresolution      = desktop
windowresolution    = 1280x960
viewport_resolution = fit
window_position     = auto
window_decorations  = true
transparency        = 0
host_rate           = auto
vsync               = true
vsync_skip          = 7000
presentation_mode   = auto
output              = opengl
texture_renderer    = auto
waitonerror         = true
priority            = auto auto
mute_when_inactive  = false
pause_when_inactive = false
mapperfile          = mapper-sdl2-0.80.0-alpha.map
screensaver         = auto

[mouse]
mouse_capture             = seamless
mouse_middle_release      = true
mouse_multi_display_aware = true
mouse_sensitivity         = 100
mouse_raw_input           = true
dos_mouse_driver          = true
dos_mouse_immediate       = false
ps2_mouse_model           = intellimouse
com_mouse_model           = wheel+msm
```