@johnnovak 2023-05-03
I've skimmed through it and nothing jumps out as being "wrong". Ultimately, I find GUI / windowing / eventing related code very hard to meaningfully review—for me, pretty much the only way is to get my hands dirty and start playing with it, trace log stuff, etc. I won't do that now, I don't have the time.

I trust @GranMinigun has a vision of where he wants to end up with all this, and I'm quite sure he has the correct overall ideas about the whole SDL cleanup / graphics pipeline optimisation stuff. So as long as improvements are made and we're truding towards that goal, I'm happy 🎉 

I'm a lot more concerned about other aspects: **testing** and our **merging strategy**.

### Testing 

I think we really should create a testing matrix for this stuff (vertical axis: features / use-cases, horizontal axis: OS + `output` combos). I can help create such a thing (e.g. a Google spreadsheet, unless better ideas), and then we should test _the living daylight_ out of this branch and all future SDL PRs in a consistent, reproducible manner as a team effort. E.g. I can personally test on Windows 10 and an M1 Mac, and only on the Mac with an external monitor, and that's it.

### Merging strategy

Clearly, we don't want this branch to get stale, or not touch anything SDL and windowing-related until our favourite Russian has fully completed his arduous quest in 6-12 months from now... But what I'm also **_very strongly against_** is to merge this now before 0.81.0 as that would overload the release and potentially push it out into the second half of the year. This is a high risk change with the potential to blow up in our face, and there _will_ be regressions and weird edge cases! I don't want to skimp on quality and throw this over the fence to our users half-tested for "community testing"—to me stability is non-negotiable. If the plane did not pass all the tests, the flight gets cancelled—no ifs and buts.

My suggestion is as follows:

1. Focus on completing 0.81.0 quickly without merging this yet. We by all means should finish the AR-correct screenshot stuff and the removal/rework of the "pixel-perfect" mode, plus finish all the smaller in-flight stuff we talked about previously. The "theme" of 0.81.0 can be then "more accurate video output" or something.
2. Once 0.81.0, we should merge this into `main` and continue testing it. That should give us ample time to iron out issues for the next 6+ months until 0.82.0, and @GranMinigun will hopefully be in a position after this to raise smaller, more incremental PRs that are easier to regression test.

Well, if I had to make an executive decision right now, this would be it 😎 🚬 . Thoughts? @dreamer @kcgen