@ThomasEricB 2023-05-04
> Well, if I had to make an executive decision right now, this would be it 😎 🚬 . Thoughts? @dreamer @kcgen

My nickname might not be @dreamer, but I am a dreamer. Even though I don't know what kcgen means outside of @kcgen 's nickname, I might be that as well! Here is what I think about your merge strategy @johnnovak : first, cigarettes are bad for your health, and second, you shouldn't wear sunglasses inside!

**Being totally serious now**, great write up, I agree 90% with you. We can't merge this whole SDL rework into main just before releasing 0.81, that would be a disaster without the proper testing it requires. However, where I diverge from you @johnnovak , is that we should focus "on completing 0.81.0 quickly". I don't think we should rush to get 0.81.0 out without just to get this PR merged into main, I don't see any problem with @GranMinigun waiting a couple of months while we carefully finish the stuff for the 0.81 release. With this position, I am placing in high priority @kcgen 's, @dreamer 's and the rest of the team's mental well-being before anything. I am worried that if we get fully invested in finishing it quickly, we might burn out the entire team. 

What we should do instead is let this PR sit on the back burner until we feel comfortably with merging it to main, after 0.81. After all, we are not on a tight schedule nor with any obligation to have 0.81 out by a certain date. So let's not start stressing with it just because of this PR, ok?