@johnnovak 2023-05-04
> What we should do instead is let this PR sit on the back burner until we feel comfortably with merging it to main, after 0.81. After all, we are not on a tight schedule nor with any obligation to have 0.81 out by a certain date. So let's not start stressing with it just because of this PR, ok?

Oh, that's what I meant! "Quickly" was maybe not the best word—the main message I wanted to get across is not to overload 0.81.0 with more stuff. I take "quickly" back then; let's finish whatever we have in flight plus the video improvements at a _comfortable_ place, but let's not add anything else non-trivial to the release. We don't want slave drivers here, that's for sure 😅