@kcgen 2023-05-04
@GranMinigun - wow - for the amount of code and complexity eliminated, I'm amazed at how robust this - no doubt, you've done a lot of regression testing!

1. LOCt: :arrow_down: 
1. Complexity: :arrow_down: 
2. Startup burden: :arrow_down: 
 - baseline (~200ms): 
![2023-05-03_17-11](https://user-images.githubusercontent.com/1557255/236086726-d2b03415-66a7-41bf-a31e-b5526a142a0c.png)
 - PR (~150ms): ![2023-05-03_17-08](https://user-images.githubusercontent.com/1557255/236086397-1c08e0b5-e591-4282-8c06-0dc3d3521188.png)

1. Performance: :arrow_up:, and specifically on bandwidth-limited hosts:
 - Pi4: QTD 320x200 using double-scanned VGA: 16 -> 18.4 fps
 - x86-64: QTD 640x480 (VESA) window-mode: 76.8 -> 84.2 fps
 - x86-64: QTD 640x480 (VESA) fullscreen 4k: 55.6 -> 64.2 fps
 - (single-scan performance is on-par because it's already pretty lean)

And finally scaled texture-drawn mapper! 

![2023-05-03_17-16](https://user-images.githubusercontent.com/1557255/236086903-c45751ec-0dfa-45f5-9408-3a22d4afbe6f.png)

That's just initial testing. Plan to start doing systematic testing like @johnnovak suggested, if we can draw up a table of functionality.