@johnnovak 2023-05-04
> That's just initial testing. Plan to start doing systematic testing like @johnnovak suggested, if we can draw up a table of functionality.

Any idea for a distributed spreadsheet we all have access to? Google Sheets is my obvious answer, but happy for other/better suggestions. Just keep it low barrier & low effort.