@kcgen 2023-05-04
> Google Sheets is my obvious answer

Works for me. @GranMinigun , is it accessible and OK for you too?